+++ .prettierrc
@@ -8,7 +8,7 @@
     "^react$",
     "<THIRD_PARTY_MODULES>",
     "^@/libs(.*)$",
-    "^@/services(.*)$",
+    "^@/api(.*)$",
     "^@/providers(.*)$",
     "^@/components(.*)$",
     "^@/hooks(.*)$",
