+++ src/api/index.ts
@@ -1,0 +1,4 @@
+export * from "./authApi";
+export * from "./usersApi";
+export * from "./charactersApi";
+export * from "./familiesApi";
