+++ src/app/(private)/onboarding/family/join/usePatchFamily.ts
@@ -1,12 +1,12 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
-import { usersApi } from "@/services";
 import { useUserData } from "@/hooks/useUserData";
 import {
   FamilyCodeFormType,
   FamilyCodeSchema,
 } from "@/schemas/FamilyCodeSchema";
+import { usersApi } from "@/api";
 
 export const usePatchFamily = () => {
   const { watch, register, formState } = useForm<FamilyCodeFormType>({
