+++ src/components/LogoutButton.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useMutation } from "@tanstack/react-query";
-import { authApi } from "@/services";
 import { useHandleAuth } from "@/hooks/useHandleAuth";
+import { authApi } from "@/api";
 
 export default function LogoutButton() {
   const { handleLogout } = useHandleAuth();
