+++ src/hooks/useUserData.ts
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { authApi, usersApi } from "@/services";
+import { authApi, usersApi } from "@/api";
 import { customToast } from "@/components";
 import queryKeys from "@/constants/queryKeys";
 import { useHandleAuth } from "./useHandleAuth";
