--- src/services/index.ts
@@ -1,3 +1,0 @@
-export * from "./auth";
-export * from "./users";
-export * from "./characters";
